master

+++ MECPerfSchema.sql
@@ -25,7 +25,7 @@ DROP TABLE IF EXISTS `BandwidthMeasure`;
 CREATE TABLE `BandwidthMeasure` (
   `id` int(11) NOT NULL,
   `sub_id` int(11) NOT NULL,
-  `nanoTimes` bigint(22) unsigned DEFAULT NULL,
+  `nanoTimes` int(22) unsigned DEFAULT NULL,
   `kBytes` decimal(10,4) unsigned DEFAULT NULL,
   PRIMARY KEY (`id`,`sub_id`),
   CONSTRAINT `fk_Test_BandwidthMeasure_id` FOREIGN KEY (`id`) REFERENCES `Test` (`ID`) ON DELETE CASCADE ON UPDATE NO ACTION
